requirements.txt
pandas~=2.2.2
matplotlib~=3.9.2
numpy~=1.26.4
scipy~=1.13.1
scikit-learn~=1.5.1
scikit-optimize==0.10.2
ydata-synthetic==1.4.0
joblib~=1.4.2
tqdm~=4.66.5
streamlit~=1.38.0
PyYAML~=6.0.2
streamlit-authenticator
pymongo~=4.8.0
requests~=2.30.0
plotly~=5.24.0